+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Showcase] Improve Fastlane alpha build notifications
 - [Library] Change type aliases for X and Y offsets of elevations tokens
+- [Library] Split raw, semantic and components tokens definitions of values, composites and aliases
 - [Library] Update border semantic tokens values ([#106](https://github.com/Orange-OpenSource/ouds-ios/issues/106))
 - [Showcase] Add fake components for demo and tokens tests
 - [Library] Remove spread value for elevation tokens